move DeprecationIssue into x-pack core and LoggingDeprecationAccumulationHandler into common namespace
