+++ x-pack/plugin/core/src/main/java/org/elasticsearch/xpack/core/deprecation/DeprecationIssue.java
@@ -4,7 +4,7 @@
  * 2.0; you may not use this file except in compliance with the Elastic License
  * 2.0.
  */
-package org.elasticsearch.xpack.deprecation;
+package org.elasticsearch.xpack.core.deprecation;
 
 import org.elasticsearch.Version;
 import org.elasticsearch.core.Nullable;
